# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/expense_tax_lines:
    get:
      tags:
        - ExpenseTaxLines
      summary: List ExpenseTaxLines
      description: Get a paginated list of ExpenseTaxLines with optional filtering and sorting
      operationId: listExpenseTaxLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseTaxLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ExpenseTaxLines
      summary: Create ExpenseTaxLine
      description: Create a new ExpenseTaxLine
      operationId: createExpenseTaxLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseTaxLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseTaxLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ExpenseTaxLine:
      type: object
      required:
        - id
        - expense_id
        - basis
        - account_id
        - rate
        - tax_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        expense_id:
          type: string
          format: uuid
        basis:
          type: string
        account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
        rate:
          type: string
        tax_amount:
          type: string
        metadata:
          type: object
...
    CreateExpenseTaxLineRequest:
      type: object
      required:
        - expense_id
        - basis
        - account_id
        - rate
        - tax_amount
        - metadata
      properties:
        expense_id:
          type: string
          format: uuid
        basis:
          type: string
        account_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
        rate:
          type: string
        tax_amount:
          type: string
        metadata:
          type: object
...
    ExpenseTaxLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseTaxLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []